+++ crates/rust-client/src/store/note_record/input_note_record/states/committed.rs
@@ -1,7 +1,7 @@
 use alloc::string::ToString;
 
 use miden_objects::Word;
-use miden_objects::block::BlockHeader;
+use miden_objects::block::{BlockHeader, BlockNumber};
 use miden_objects::note::{NoteId, NoteInclusionProof, NoteMetadata};
 use miden_objects::transaction::TransactionId;
 
@@ -42,7 +42,7 @@ impl NoteStateHandler for CommittedNoteState {
 
     fn consumed_externally(
         &self,
-        nullifier_block_height: u32,
+        nullifier_block_height: BlockNumber,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(Some(ConsumedExternalNoteState { nullifier_block_height }.into()))
     }
@@ -86,7 +86,7 @@ impl NoteStateHandler for CommittedNoteState {
     fn transaction_committed(
         &self,
         _transaction_id: TransactionId,
-        _block_height: u32,
+        _block_height: BlockNumber,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Err(NoteRecordError::InvalidStateTransition(
             "Only processing notes can be committed in a local transaction".to_string(),