+++ crates/web-client/test/global.test.d.ts
@@ -32,6 +32,7 @@ import {
   NoteExecutionHint,
   NoteExecutionMode,
   NoteFilter,
+  NoteFile,
   NoteFilterTypes,
   NoteId,
   NoteIdAndArgs,
@@ -118,6 +119,7 @@ declare global {
     NoteExecutionHint: typeof NoteExecutionHint;
     NoteExecutionMode: typeof NoteExecutionMode;
     NoteFilter: typeof NoteFilter;
+    NoteFile: typeof NoteFile;
     NoteFilterTypes: typeof NoteFilterTypes;
     NoteId: typeof NoteId;
     NoteIdAndArgs: typeof NoteIdAndArgs;